move user specific db operations to db/users dir
