+++ {{cookiecutter.project_slug}}/backend/app/core/auth.py
@@ -2,8 +2,10 @@ import jwt
 from fastapi import Depends, HTTPException, status
 from jwt import PyJWTError
 
-from app.db import models, schemas, session
-from app.db.crud import get_user_by_email, create_user
+from app.db import session
+from app.db.users.crud import get_user_by_email, create_user
+from app.db.users.models import User
+from app.db.users.schemas import TokenData, UserCreate
 from app.core import security
 
 
@@ -23,7 +25,7 @@ async def get_current_user(
         if email is None:
             raise credentials_exception
         permissions: str = payload.get("permissions")
-        token_data = schemas.TokenData(email=email, permissions=permissions)
+        token_data = TokenData(email=email, permissions=permissions)
     except PyJWTError:
         raise credentials_exception
     user = get_user_by_email(db, token_data.email)
@@ -33,7 +35,7 @@ async def get_current_user(
 
 
 async def get_current_active_user(
-    current_user: models.User = Depends(get_current_user),
+    current_user: User = Depends(get_current_user),
 ):
     if not current_user.is_active:
         raise HTTPException(status_code=400, detail="Inactive user")
@@ -41,8 +43,8 @@ async def get_current_active_user(
 
 
 async def get_current_active_superuser(
-    current_user: models.User = Depends(get_current_user),
-) -> models.User:
+    current_user: User = Depends(get_current_user),
+) -> User:
     if not current_user.is_superuser:
         raise HTTPException(
             status_code=403, detail="The user doesn't have enough privileges"
@@ -63,13 +65,9 @@ def sign_up_new_user(db, email: str, password: str):
     user = get_user_by_email(db, email)
     if user:
         return False  # User already exists
-    new_user = create_user(
-        db,
-        schemas.UserCreate(
-            email=email,
-            password=password,
-            is_active=True,
-            is_superuser=False,
-        ),
-    )
-    return new_user
+    return create_user(db, UserCreate(
+        email=email,
+        password=password,
+        is_active=True,
+        is_superuser=False,
+    ))
