+++ {{cookiecutter.project_slug}}/backend/app/db/users/schemas.py
