+++ {{cookiecutter.project_slug}}/backend/app/initial_data.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 
 from app.db.session import get_db
-from app.db.crud import create_user
-from app.db.schemas import UserCreate
+from app.db.users.crud import create_user
+from app.db.users.schemas import UserCreate
 from app.db.session import SessionLocal
