+++ {{cookiecutter.project_slug}}/backend/conftest.py
@@ -7,7 +7,7 @@ import typing as t
 
 from app.core import config, security
 from app.db.session import Base, get_db
-from app.db import models
+from app.db.users.models import User
 from app.main import app
 
 
@@ -100,12 +100,12 @@ def get_password_hash() -> str:
 
 
 @pytest.fixture
-def test_user(test_db) -> models.User:
+def test_user(test_db) -> User:
     """
     Make a test user in the database
     """
 
-    user = models.User(
+    user = User(
         email="fake@email.com",
         hashed_password=get_password_hash(),
         is_active=True,
@@ -116,12 +116,12 @@ def test_user(test_db) -> models.User:
 
 
 @pytest.fixture
-def test_superuser(test_db) -> models.User:
+def test_superuser(test_db) -> User:
     """
     Superuser for testing
     """
 
-    user = models.User(
+    user = User(
         email="fakeadmin@email.com",
         hashed_password=get_password_hash(),
         is_superuser=True,
